fix: test issues from refactor
