+++ tests/layers/test_categorical_calibrator.py
@@ -6,7 +6,7 @@ import torch
 from pytorch_lattice.enums import CategoricalCalibratorInit
 from pytorch_lattice.layers import CategoricalCalibrator
 
-from ..utils import train_calibrated_module
+from ..test_utils import train_calibrated_module
 
 
 @pytest.mark.parametrize(
